Fix ruff/isort/black/mypy findings; first real tools/check.sh run
Lint and mypy had never actually been run locally (not installed in
the conda env), so this is the first time tools/check.sh (what CI
runs) executed end to end. Fixes 235 ruff errors, reformats 45 files
under black, and resolves all 173 mypy errors.

Most of the mypy fixes are annotation corrections (NPFArray/NPIArray
were hardcoded to float32/int64 but the codebase genuinely mixes
precisions; several functions had return-type/arity annotations that
didn't match their actual return values) and a couple of test-double
Protocols (PdfFitter, WalkSimulator) so fakes don't need to subclass
production fitters/simulators.

A handful are real bugs surfaced along the way, all in code paths
nothing currently exercises:
- visualizer.py used vtk.vtkTimerCallbackCamera, which doesn't exist
  in VTK (it's our own class) - would AttributeError under
  animation=True.
- draw_trajectory_points called two helpers without their required
  kwargs - would crash on first call; never actually called anywhere.
- utils.create_box_mask called BoundingBox.inside() with one arg
  instead of x/y/z - never called anywhere either.
- read_pnm_ext_linklist's return type claimed both arrays were int32;
  the second (throat lengths) is float32.
- skip_struct_main_part and get_up_down were missing @staticmethod;
  harmless only because of how they happened to be called.

pytest: 88 passed, same as before.

+++ tests/test_trap_distr_builder.py
@@ -38,8 +38,8 @@ def _table_row(gas: str, mu: float) -> TableIIIRow:
 
 def test_table_iii_event_metrics_average_per_trajectory_probability() -> None:
     sequences = [
-        TrapSequence(np.ones(2), np.array([0.0, 1.0])),
-        TrapSequence(np.ones(4), np.array([0.0, 0.0, 0.0, 1.0])),
+        TrapSequence(np.ones(2, dtype=int), np.array([0.0, 1.0])),
+        TrapSequence(np.ones(4, dtype=int), np.array([0.0, 0.0, 0.0, 1.0])),
     ]
 
     summary = summarize_trap_events(sequences)
